+++ docs/changelog.rst
@@ -48,6 +48,8 @@ Detailed list of changes
 
 - A new option :opt:`menu_map` that allows adding entries to the global menubar on macOS (:disc:`6680`)
 
+- A new :doc:`escape code <pointer-shapes>` that can be used by programs running in the terminal to change the shape of the mouse pointer (:iss:`6711`)
+
 - A new mouse action ``mouse_selection word_and_line_from_point`` to select the current word under the mouse cursor and extend to end of line (:pull:`6663`)
 
 - Allow using the full range of standard mouse cursor shapes when customizing the mouse cursor